Changed top level plugin file name. Discovered several references to "template" they were changed to cb-pdp_schedule.
note: when changing template to cb-pdp_schedule I got Wordpress critical error message.
As I worked through the errors, I commented out things is the _constructor function in
class-init and worked my way down the function call path, commenting and uncommenting
lines to find the problem. In the end the code is working. Perhaps the old code was
cached somehow?